Bug fixes  - AV_Decision: casting double to int  - Infection: no breaks in switch statement, recovery period < 0 handled incorrectly  - stats: underflow when calculating stdev
git-svn-id: https://redmine.hindsight.psc.edu/svn/fred/trunk/FRED@225 b93b844c-2574-459d-9dfb-c4b8cb2af1a1

+++ AV_Decisions.cc
@@ -48,7 +48,7 @@ Decision(p){
 int AV_Decision_Give_to_Sympt::evaluate(Person* person, int strain, int current_day){
   AV_Manager *avm = dynamic_cast < AV_Manager* > ( policy->get_manager() );
   Antiviral* av = avm->get_current_av();
-  int percentage = av->get_percent_symptomatics();
+  double percentage = av->get_percent_symptomatics();
   if(person->get_health()->get_strain_status(strain) == 'I'){
     person->get_health()->flip_checked_for_av(strain);
     double r = RANDOM(); // This is now a probability <=1.0;
